Python/CertificateUtil/pyproject.toml
[tool.poetry]
name = "certificateutil"
version = "0.1.0"
description = ""
authors = ["kero1976 <kero1976@outlook.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyopenssl = "^24.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
